requirements.txt
Package            Version
------------------ -------
aiohttp            3.8.3
aiosignal          1.3.1
async-timeout      4.0.2
attrs              22.2.0
autopep8           2.0.1
charset-normalizer 2.1.1
discord            2.1.0
discord.py         2.1.0
frozenlist         1.3.3
idna               3.4
multidict          6.0.4
pip                23.0
pycodestyle        2.10.0
python-dotenv      0.21.1
setuptools         65.6.3
wheel              0.38.4
yarl               1.8.2
